Refractored DinosaurController class
