+++ src/main/java/com/dinosaur/dinosaurexploder/controller/core/CollisionRegistry.java
@@ -1,0 +1,18 @@
+package com.dinosaur.dinosaurexploder.controller.core;
+
+import com.dinosaur.dinosaurexploder.controller.core.collisions.CollisionHandlerInterface;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CollisionRegistry {
+    private final List<CollisionHandlerInterface> collisions = new ArrayList<>();
+
+    public void addCollision(CollisionHandlerInterface collision) {
+        collisions.add(collision);
+    }
+
+    public void registerAll() {
+        collisions.forEach(CollisionHandlerInterface::register);
+    }
+}
