+++ src/main/java/com/dinosaur/dinosaurexploder/controller/core/collisions/CollisionHandlerInterface.java
@@ -1,0 +1,5 @@
+package com.dinosaur.dinosaurexploder.controller.core.collisions;
+
+public interface CollisionHandlerInterface {
+    void register();
+}
